requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pytesseract==0.3.13
sqlparse==0.5.1
tzdata==2024.2
